+++ portal-impl/src/META-INF/portal-log4j.xml
@@ -318,7 +318,7 @@
 		<priority value="INFO" />
 	</category>
 
-	<category name="com.liferay.portal.kernel.util.JavaProps">
+	<category name="com.liferay.portal.kernel.util.JavaDetector">
 		<priority value="ERROR" />
 	</category>
