+++ portal-impl/src/com/liferay/portal/util/InitUtil.java
@@ -24,7 +24,7 @@ import com.liferay.portal.kernel.dao.db.DBFactoryUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataSourceFactoryUtil;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.JavaProps;
+import com.liferay.portal.kernel.util.JavaDetector;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.StringPool;
@@ -126,7 +126,7 @@ public class InitUtil {
 
 		// Java properties
 
-		JavaProps.isJDK5();
+		JavaDetector.isJDK5();
 
 		// ROME
